lazy-load every SPA page
