+++ static/js/spa/components/LoadingSpinner.js
@@ -1,0 +1,6 @@
+import { html } from 'htm/preact';
+
+export function LoadingSpinner() {
+    return html`<h2 className="m-1">Loading...</h2>`;
+}
+
