+++ static/js/spa/mps/components/AddStreamPage.js
@@ -2,6 +2,6 @@ import { html } from 'htm/preact';
 
 import { EditStreamCard } from './EditStreamCard.js';
 
-export function AddStreamPage() {
+export default function AddStreamPage() {
   return html`<${EditStreamCard} name=".add" newStream />`;
 }
