+++ static/js/spa/mps/components/EditStreamPage.js
@@ -3,7 +3,7 @@ import { useParams } from "wouter-preact";
 
 import { EditStreamCard } from './EditStreamCard.js';
 
-export function EditStreamPage() {
+export default function EditStreamPage() {
   const {mps_name} = useParams();
   return html`<${EditStreamCard} name="${mps_name}" />`;
 }
